+++ edgeimpulse/experimental/impulse/__init__.py
@@ -1,0 +1,9 @@
+"""Module to build and deploy impulses in Edge Impulse."""
+
+from edgeimpulse.impulse._functions.build import (
+    build,
+)
+
+__all__ = [
+    "build",
+]