+++ edgeimpulse/impulse/__init__.py
@@ -1,0 +1,1 @@
+"""Use this module to build and deploy impulses from Edge Impulse."""